upd Project wsd file
