+++ src/main/java/com/geekbrains/geekspring/wsdproducts/generation/GetProductsRequest.java
@@ -2,11 +2,11 @@
 // This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.7 
 // See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
 // Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2021.03.21 at 12:59:47 PM MSK 
+// Generated on: 2021.03.21 at 01:22:51 PM MSK 
 //
 
 
-package geekbrains_com.geekspring.wsdproducts.generation;
+package com.geekbrains.geekspring.wsdproducts.generation;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
@@ -25,7 +25,7 @@ import javax.xml.bind.annotation.XmlType;
  *   &lt;complexContent>
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
  *       &lt;sequence>
- *         &lt;element name="product" type="{http:www//geekbrains.com/geekspring/wsdproducts/generation}product"/>
+ *         &lt;element name="name" type="{http://www.w3.org/2001/XMLSchema}string"/>
  *       &lt;/sequence>
  *     &lt;/restriction>
  *   &lt;/complexContent>
@@ -36,36 +36,36 @@ import javax.xml.bind.annotation.XmlType;
  */
 @XmlAccessorType(XmlAccessType.FIELD)
 @XmlType(name = "", propOrder = {
-    "product"
+    "name"
 })
-@XmlRootElement(name = "getProductsResponse")
-public class GetProductsResponse {
+@XmlRootElement(name = "getProductsRequest")
+public class GetProductsRequest {
 
     @XmlElement(required = true)
-    protected Product product;
+    protected String name;
 
     /**
-     * Gets the value of the product property.
+     * Gets the value of the name property.
      * 
      * @return
      *     possible object is
-     *     {@link Product }
+     *     {@link String }
      *     
      */
-    public Product getProduct() {
-        return product;
+    public String getName() {
+        return name;
     }
 
     /**
-     * Sets the value of the product property.
+     * Sets the value of the name property.
      * 
      * @param value
      *     allowed object is
-     *     {@link Product }
+     *     {@link String }
      *     
      */
-    public void setProduct(Product value) {
-        this.product = value;
+    public void setName(String value) {
+        this.name = value;
     }
 
 }
