+++ src/main/java/com/geekbrains/geekspring/wsdproducts/generation/GetProductsResponse.java
@@ -2,11 +2,11 @@
 // This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.7 
 // See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
 // Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2021.03.21 at 12:59:47 PM MSK 
+// Generated on: 2021.03.21 at 01:22:51 PM MSK 
 //
 
 
-package geekbrains_com.geekspring.wsdproducts.generation;
+package com.geekbrains.geekspring.wsdproducts.generation;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
@@ -25,7 +25,7 @@ import javax.xml.bind.annotation.XmlType;
  *   &lt;complexContent>
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
  *       &lt;sequence>
- *         &lt;element name="name" type="{http://www.w3.org/2001/XMLSchema}string"/>
+ *         &lt;element name="product" type="{http:www.geekbrains.com/geekspring/wsdproducts/generation}product"/>
  *       &lt;/sequence>
  *     &lt;/restriction>
  *   &lt;/complexContent>
@@ -36,36 +36,36 @@ import javax.xml.bind.annotation.XmlType;
  */
 @XmlAccessorType(XmlAccessType.FIELD)
 @XmlType(name = "", propOrder = {
-    "name"
+    "product"
 })
-@XmlRootElement(name = "getProductsRequest")
-public class GetProductsRequest {
+@XmlRootElement(name = "getProductsResponse")
+public class GetProductsResponse {
 
     @XmlElement(required = true)
-    protected String name;
+    protected Product product;
 
     /**
-     * Gets the value of the name property.
+     * Gets the value of the product property.
      * 
      * @return
      *     possible object is
-     *     {@link String }
+     *     {@link Product }
      *     
      */
-    public String getName() {
-        return name;
+    public Product getProduct() {
+        return product;
     }
 
     /**
-     * Sets the value of the name property.
+     * Sets the value of the product property.
      * 
      * @param value
      *     allowed object is
-     *     {@link String }
+     *     {@link Product }
      *     
      */
-    public void setName(String value) {
-        this.name = value;
+    public void setProduct(Product value) {
+        this.product = value;
     }
 
 }
