+++ src/main/java/com/geekbrains/geekspring/wsdproducts/generation/ObjectFactory.java
@@ -2,11 +2,11 @@
 // This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.7 
 // See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
 // Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2021.03.21 at 12:59:47 PM MSK 
+// Generated on: 2021.03.21 at 01:22:51 PM MSK 
 //
 
 
-package geekbrains_com.geekspring.wsdproducts.generation;
+package com.geekbrains.geekspring.wsdproducts.generation;
 
 import javax.xml.bind.annotation.XmlRegistry;
 
@@ -14,7 +14,7 @@ import javax.xml.bind.annotation.XmlRegistry;
 /**
  * This object contains factory methods for each 
  * Java content interface and Java element interface 
- * generated in the geekbrains_com.geekspring.wsdproducts.generation package. 
+ * generated in the com.geekbrains.geekspring.wsdproducts.generation package. 
  * <p>An ObjectFactory allows you to programatically 
  * construct new instances of the Java representation 
  * for XML content. The Java representation of XML 
@@ -30,7 +30,7 @@ public class ObjectFactory {
 
 
     /**
-     * Create a new ObjectFactory that can be used to create new instances of schema derived classes for package: geekbrains_com.geekspring.wsdproducts.generation
+     * Create a new ObjectFactory that can be used to create new instances of schema derived classes for package: com.geekbrains.geekspring.wsdproducts.generation
      * 
      */
     public ObjectFactory() {
