+++ src/main/java/com/geekbrains/geekspring/wsdproducts/generation/package-info.java
@@ -2,8 +2,8 @@
 // This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.7 
 // See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
 // Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2021.03.21 at 12:59:47 PM MSK 
+// Generated on: 2021.03.21 at 01:22:51 PM MSK 
 //
 
-@javax.xml.bind.annotation.XmlSchema(namespace = "http:www//geekbrains.com/geekspring/wsdproducts/generation", elementFormDefault = javax.xml.bind.annotation.XmlNsForm.QUALIFIED)
-package geekbrains_com.geekspring.wsdproducts.generation;
+@javax.xml.bind.annotation.XmlSchema(namespace = "http:www.geekbrains.com/geekspring/wsdproducts/generation", elementFormDefault = javax.xml.bind.annotation.XmlNsForm.QUALIFIED)
+package com.geekbrains.geekspring.wsdproducts.generation;
